Streamlines reloading logic to use ScheduledExecutorService
The next big red flag to tackle was the huge loop logic in the
ProjectDataReloader. If you have to handle periodically repeating tasks
that are run in Threads, it's a pretty sure bet that Java's executor
classes have done it better already.

After this refactor, all work of scheduling is done by the
ScheduledExecutorService, and the ProjectDataReloader is only looping in
order to stop.


In order to take advantage of the data structure, we needed to be able
to express our reload policies in terms of time, not "reload counts."
Remember that thing where I said that having the "reload policy" would
give us one place to change the logic around reloading?  This is where
that pays off.

Removed the reloadsCounter altogether, since we don't care anymore.