Eliminates all Threads, waits, and looping from ProjectDataReloader.
The final step is to simplify the start() and stop() methods of the PDR
class by holding on to the ScheduledExecutorService instead of the
"stopped" boolean.  By surfacing the former, we can perform the shutdown
in the stop() method.

That way, there's no need to listen for/care about the concept of
"stopping" in the start() method.

Also tightened visibility and removed now-unnecessary class members.