Created named "loader" classes to hold loading-related code.
This fixes the "single responsibility principle" violation in
ProjectDataLoader.

ProjectDataLoader used to be responsible for:
- scheduling which types of data to load
- knowing how to load those data types

To fix that, we pull the "how do I load this?" code into a class of its
own; and rather than have one giant "how do I load everything" class, we
have one class *per data type*.  I.e. a ProjectDetailsLoader, a
LoginStatisticsLoader, and a LastUpdateTimeLoader.

This also gets rid of the (kind of ugly) inline definition of our
runnables.