Removes subtypes of ProjectDataReloader
This final step cleans up the type hierarchy of ProjectDataReloader.
The only difference between the Live and Static project reloaders was in
the list of DataReloader objects being built.  We can accommodate that
difference without having named subtypes.

This is an example of "favoring composition over inheritance."  This
way, a "live" project's data reloader *has a* list that contains three
reloaders, rather than *being a* LiveProjectDataReloader; the "static"
project *has a* list with one reloader, rather than *being a*
StaticProjectDataReloader.